Add ndt7 spec and sync the implementation
Most likely always using float64 is going to save us headaches
in the event in which we're dealing with 32 bit clients.

There are still 32 bit devices around (i.e. Android and iOS devices)
so I believe that this is a valid concern.

Incidentally, this move remove any overflow concern when adding
up bytes because there is no overflow with float64 just a loss of
precision when you use more than 53 bits for the value.

+++ ndt7/client.go
@@ -33,7 +33,7 @@ func (cl Client) Download() error {
 	conn.SetReadLimit(MinMaxMessageSize)
 	defer conn.Close()
 	t0 := time.Now()
-	num := int64(0)
+	num := float64(0.0)
 	ticker := time.NewTicker(MinMeasurementInterval)
 	log.Info("Starting download")
 	for {
@@ -59,7 +59,7 @@ func (cl Client) Download() error {
 			}
 			break
 		}
-		num += int64(len(mdata))
+		num += float64(len(mdata))
 		if mtype == websocket.TextMessage {
 			// Unmarshal to verify that this message is correct JSON but do not
 			// otherwise process the message's content.
